requirements.txt
adal==1.2.1
aiohttp==3.6.2
aiounittest==1.4.0
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
Babel==2.7.0
botbuilder-community-dialogs-prompts==0.1.3
botbuilder-core==4.9.1
botbuilder-dialogs==4.9.1
botbuilder-schema==4.9.1
botframework-connector==4.9.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
datedelta==1.3
emoji==0.5.4
grapheme==0.6.0
idna==2.8
isodate==0.6.0
jsonpickle==1.2
msal==1.2.0
msrest==0.6.10
multidict==4.7.6
multipledispatch==0.6.0
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.5.3
python-dateutil==2.8.1
pytz==2020.1
recognizers-text==1.0.2a2
recognizers-text-choice==1.0.2a2
recognizers-text-date-time==1.0.2a2
recognizers-text-number==1.0.2a2
recognizers-text-number-with-unit==1.0.2a2
recognizers-text-sequence==1.0.2a2
recognizers-text-suite==1.0.2a2
regex==2019.8.19
requests==2.22.0
requests-oauthlib==1.3.0
six==1.15.0
ujson==3.0.0
urllib3==1.25.9
wrapt==1.12.1
yarl==1.4.2
